backend/requirements.txt
# FastAPI
fastapi==0.115.8
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.1

# Settings & Environment
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Validation
email-validator==2.1.1

# Utilities
python-multipart==0.0.20
